TimeDisplay Component Added
